requirements.txt
streamlit
python_dotenv
langchain
openai

transformers
torch
accelerate 
bitsandbytes
scipy